requirements.txt
pdfplumber==0.10.4
sentence-transformers==2.7.0
faiss-cpu==1.8.0
requests==2.32.3